Remove user examples and Shiny code from "ready-to-go" branch.
Both the user examples and the Shiny application are works in progress to one degree or another, so I've left them in the `kitchen-sink` branch but am removing them from the `ready-to-go` branch that we're requesting be pulled into the ropensci package. If we want to work more on these in the future, they can be found in the `kitchen_sink` branch.