addressed product="datums" without durations
datums product does not in fact _Require_ begin_date & end_date, and coops_search() does not test !is.null() on those parameters.

Note also that station_name is listed as "numeric", but it clearly can be character.  For product="currents", it _must_ be character, as the station_name values for all stations with currents data start with an alphabetic character, not a numeral.  See the example on the man page:
'''coops_search(station_name = "ps0401", begin_date = 20151221, end_date = 20151222,
 product = "currents")'''